# lib/api-zod/src/generated/types/computeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkDetails } from './workDetails';

export interface ComputeInput {
  works: WorkDetails[];
}