# lib/api-zod/src/generated/types/contractor.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Contractor {
  id: number;
  /** @nullable */
  sno?: number | null;
  name: string;
  address: string;
  /** @nullable */
  contactNo?: string | null;
  createdAt?: string;
}