# lib/api-zod/src/generated/types/contractorInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ContractorInput {
  /** @nullable */
  sno?: number | null;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  address: string;
  /** @nullable */
  contactNo?: string | null;
}