# lib/api-zod/src/generated/types/diprEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DiprEntry {
  srNo: number;
  newspaperName: string;
  editionName: string;
  advtSize: string;
  sqCm: number;
  releaseDate: string;
  rate: number;
  amount: number;
}