# lib/api-zod/src/generated/types/generatedDocument.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedDocument {
  type: string;
  title: string;
  downloadToken: string;
  filename: string;
}