# lib/api-zod/src/generated/types/nitSessionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallanEntry } from './challanEntry';
import type { DiprPublication } from './diprPublication';
import type { NitDetails } from './nitDetails';
import type { WorkDetails } from './workDetails';

export interface NitSessionInput {
  title: string;
  nitDetails: NitDetails;
  works: WorkDetails[];
  challanEntries?: ChallanEntry[];
  diprPublication?: DiprPublication | null;
}