# lib/api-zod/src/generated/types/nitSessionSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PWD NIT Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkComputation } from './workComputation';

export interface NitSessionSummary {
  sessionId: number;
  totalWorks: number;
  acceptedWorks: number;
  cancelledWorks: number;
  totalBidValue: number;
  totalStampDuty: number;
  totalAps: number;
  workComputations: WorkComputation[];
}